Rename parser directory to endpoint-parser
